Fix OrderRequest test on getAllAsynch
Unlike it is made in the other logic class, in the order logic, the GetResponseAsync method that has OrderRequest has parameter, don't test if this parameter is null to affect it the default. This orderRequest parameter is always overwrite by the default, so no filter on orders pass to the getAllAsync is used.